Add report for subscriber lists with ability to limit total number
Allows for reporting on subscriber counts for single page notifications.
It is scoped to subscriber lists with content_ids, which will limit the
output to just single page notification subscriber lists.

It defaults a limit to the top 25 subscriber lists, however this can be
overidden by passing an limit parameter.

Finally to support tests I've extended the factory for the
single_page_subscription trait to allow lists to be created without a
clash of URLs, and for subscribers to be associated (randomly allocated
between 0 and 3).