Delete nullified accounts from account API
The only use of the GOV.UK account on GOV.UK is for email notifications.
Therefore, to minimise us hoarding personal data we should delete a
user's GOV.UK account at the same time it's nullified in this app to
minimise data hoarding.

Update the nullify worker that removes old accounts with no active
subscriptions to also call account API if that subscriber record has
a `govuk_account_id`.