Explicitly copy from template1
Necessary for when this is called by the ci agent.
CI Agent runs without the permission to install the
uuid-ossp extension. You can manually create the extension
in template1 and have it copied into all the new
databases by default, but rails ignores this unless you
explicitly tell it to use template1.