Email subscriptions for document collections
When a user signs up to alerts on a document collection, research
has shown an expectation of being notified when changes are made to
documents within that collection.

The relationship between a document collection and the documents it
contains is stored in the links attribute sent to this application by
email alert service. See [1]

The current logic that matches subscriber lists to a content change
matches on content ID AND tags or links. This code was added in [2] when we began
supporting single page email notifications. I've followed that pattern here. eg
if the ContentChange has tags as well as a content id, the SubscriberListQuery will
only match it to a subscriber list if it matches on the tags AND the content ID.

In the case of a user being subscribed to a single page AND a collection
that includes that page, they will recieve two emails.

[1] https://github.com/alphagov/email-alert-service/pull/577
[2] https://github.com/alphagov/email-alert-api/pull/1640/commits/069ef4eafe87b751ce7cf352f8137b7bd99ac0fd