Remove test cases for content id
These tests create a subscriber list based on the attributes
passed in to the method calls `includes_subscriber_list` or
`excludes_subscriber_list`. And then we see if the
SubscriberListQuery sucessfully matches that list against the
content change params that it's instatiated with.

However the types of list being tested are not representative of any
SubscriberLists that this application has in its database. On prod:

SubscriberList.where.not(content_id: nil).pluck(:links)
SubscriberList.where.not(content_id: nil).pluck(:tags)

Both return an array of empty hashes.

This is because when email alert frontend sends a find_or_create query for
a content-id based subscription (aka a single page notification), it only
sends the content id. It does not send any links or tags [1].

Replacement tests will be added in subsequent commit.

[1] https://github.com/alphagov/email-alert-frontend/blob/main/app/controllers/single_page_subscriptions_controller.rb#L65-L71