Add a new test to show undesirable behaviour
When we rolled out email subscriptions based on content-id
we followed the existing matching rules. eg
If both links and document_types are present, both must match.
If both tags and document_types are present, both must match.
If both content id and document_types are present, both must match.

This commit adds a test to highlight this behaviour.

Currently this behaviour is trivial because we do not store document types
when we create content-id based subscriber lists.