BadListSubscriptionsMigrator class
In April 2023 a bug was detected in email alert frontend [1], that had resulted
in users who should have been signed up to a links based subscriber list (eg on an organisation page,
or topic page), being signed up to a content id based subscriberlist instead.

These 'bad' subscriptions have a content id with no links or tags, for content types that are supposed
to be links based. Which are these ones [2]

This (temporary) class can be called to migrate users from a bad list to the equivalent good list for
specialist topic email lists only. The task might be safe to run for all prefixes in ref [2] but some
checks should be done first to make sure there are no strange edge cases.

Implementation details

- Only migrate users to lists that are active:
When a specialist topic is archived, all of its email subscribers are bulk unsubscribed [3],
leaving the list with an active_subscription_count of 0. It is possible that in the time since
a bad lists was created for a specialist topic, the page itself has been archived.
There is no point migrating users to these lists as they won't recieve any updates.

- Don't call the SubscriberListMover if the bad list has no active subscribers:
When the bad lists were created, an email was sent to users asking them to unsubscribe
and resubscribe. So it's possible that there are some bad lists without subscribers. So
let's skip those.

[1] https://github.com/alphagov/email-alert-frontend/pull/1522
[2] https://github.com/alphagov/email-alert-frontend/blob/main/app/services/subscriber_list_params/generate_links_based_list_params_service.rb#L22-L43
[3] https://github.com/alphagov/collections-publisher/blob/24a7108e588f419892b52c4a8fa4584ea3c27496/app/services/tag_archiver.rb#L61