Remove code to output email addresses
The subscriber_list_mover outputs the email addresses of the users being
migrated to new lists. This can be removed because am pretty sure we shouldn't
be doing that from PII pov. And also now we are running tasks on the command line
instead of jenkins, that is a lot of logging which isn't really of any value.