Add task to delete bad SubscriberLists
These lists were created by accident in April 2023, the active
subscribers were migrated away from the lists by this PR:
https://github.com/alphagov/email-alert-api/pull/1948

The lists will delete automatically, but not for a year because
of the one-year retention period for ended subscriptions. In this
case, the data is confusing so worthwhile clearing up earlier.

We won't touch the Subscribers, they'll be retained for a year if
they have no active subscriptions, then cleared up automatically.