Stop sending duplicate confirmation emails
We've received a report of duplicate email confirmations being sent out
when a user has subscribed to a number of subscriptions. This change
prevents a confirmation email being sent if the user is already
subscribed and the frequency is the same.

The confirmation page in Email Alert Frontend remains unchanged.

As the CreateSubscriptionServive uses a lock on the subscriber record,
this should prevent duplicate emails being sent out for issues like
double clicks as the lock will only be released when the record has been
created.

Trello:
https://trello.com/c/hJ5up59M/2249-stop-sending-duplicate-signup-emails-if-user-clicks-confirmation-more-than-once