Add Prometheus configuration to collect new metrics
We have metrics still configured to send to StatsD
which has been deprecated. They are probably still
useful to collect, so we should start reporting
them to Prometheus instead.

We first need to register each metric[^1] with the
PrometheusExporter client[^2]. Once they are registered
we should be able to send data to Prometheus using
the `observe` method.

We currently have a custom type collector[^3] to send
metrics about travel advice and medical alerts in
a batch (and is configured as a cron job in
govuk-helm-charts). We might be able to combine
this, but for now we will keep the new configuration
separate.

Reference:
[^1] https://prometheus.io/docs/concepts/metric_types
[^2]: https://github.com/discourse/prometheus_exporter/?tab=readme-ov-file#single-process-mode
[^3]: https://github.com/discourse/prometheus_exporter/?tab=readme-ov-file#custom-type-collectors